feat: Gimmie more feature for case studies
